requirements.txt
Django==1.9.2
Markdown==2.6.5
django-filter==0.12.0
djangorestframework==3.3.2
psycopg2==2.6.1
virtualenv==13.1.2
